+++ MacPacker/CHANGELOG.md
@@ -1,6 +1,7 @@
 v0.12
 - feat: Added 'File' > 'Open...' menu to open archives when no window is open
 - core: Separated localization of changelog from app localization
+- core: Language support for Ukranian, Russian
 
 v0.11
 - feat: Use system preview instead of internal previewer